refactor(e2e): move mirror-registry tests under product
